Make useBlockEditingMode() public
