+++ packages/block-editor/src/store/test/private-selectors.js
@@ -1,23 +1,14 @@
 /**
- * WordPress dependencies
- */
-import { select } from '@wordpress/data';
-
-/**
  * Internal dependencies
  */
 import {
 	isBlockInterfaceHidden,
 	getLastInsertedBlocksClientIds,
-	getBlockEditingMode,
 	isBlockSubtreeDisabled,
 	getEnabledClientIdsTree,
 	getEnabledBlockParents,
 } from '../private-selectors';
-
-jest.mock( '@wordpress/data/src/select', () => ( {
-	select: jest.fn(),
-} ) );
+import { getBlockEditingMode } from '../selectors';
 
 describe( 'private selectors', () => {
 	describe( 'isBlockInterfaceHidden', () => {
@@ -63,7 +54,7 @@ describe( 'private selectors', () => {
 		} );
 	} );
 
-	describe( 'block editing mode selectors', () => {
+	describe( 'isBlockSubtreeDisabled', () => {
 		const baseState = {
 			settings: {},
 			blocks: {
@@ -130,264 +121,100 @@ describe( 'private selectors', () => {
 			blockEditingModes: new Map( [] ),
 		};
 
-		describe( 'getBlockEditingMode', () => {
-			const __experimentalHasContentRoleAttribute = jest.fn(
-				() => false
-			);
-			select.mockReturnValue( {
+		const __experimentalHasContentRoleAttribute = jest.fn( () => false );
+		getBlockEditingMode.registry = {
+			select: jest.fn( () => ( {
 				__experimentalHasContentRoleAttribute,
-			} );
+			} ) ),
+		};
 
-			it( 'should return default by default', () => {
-				expect(
-					getBlockEditingMode(
-						baseState,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'default' );
-			} );
-
-			it( 'should return disabled if explicitly set', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'disabled' ],
-					] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'disabled' );
-			} );
-
-			it( 'should return contentOnly if explicitly set', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[
-							'b3247f75-fd94-4fef-97f9-5bfd162cc416',
-							'contentOnly',
-						],
-					] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'contentOnly' );
-			} );
-
-			it( 'should return disabled if explicitly set on a parent', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
-					] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'disabled' );
-			} );
-
-			it( 'should return default if parent is set to contentOnly', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[
-							'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337',
-							'contentOnly',
-						],
-					] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'default' );
-			} );
-
-			it( 'should return disabled if overridden by a parent', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ '', 'disabled' ],
-						[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'default' ],
-						[ '9b9c5c3f-2e46-4f02-9e14-9fe9515b958f', 'disabled' ],
-					] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'disabled' );
-			} );
-
-			it( 'should return disabled if explicitly set on root', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [ [ '', 'disabled' ] ] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'disabled' );
-			} );
-
-			it( 'should return default if root is contentOnly', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [ [ '', 'contentOnly' ] ] ),
-				};
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'default' );
-			} );
-
-			it( 'should return disabled if parent is locked and the block has no content role', () => {
-				const state = {
-					...baseState,
-					blockListSettings: {
-						...baseState.blockListSettings,
-						'9b9c5c3f-2e46-4f02-9e14-9fe9515b958f': {
-							templateLock: 'contentOnly',
-						},
-					},
-				};
-				__experimentalHasContentRoleAttribute.mockReturnValueOnce(
-					false
-				);
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'disabled' );
-			} );
-
-			it( 'should return contentOnly if parent is locked and the block has a content role', () => {
-				const state = {
-					...baseState,
-					blockListSettings: {
-						...baseState.blockListSettings,
-						'9b9c5c3f-2e46-4f02-9e14-9fe9515b958f': {
-							templateLock: 'contentOnly',
-						},
-					},
-				};
-				__experimentalHasContentRoleAttribute.mockReturnValueOnce(
-					true
-				);
-				expect(
-					getBlockEditingMode(
-						state,
-						'b3247f75-fd94-4fef-97f9-5bfd162cc416'
-					)
-				).toBe( 'contentOnly' );
-			} );
+		it( 'should return false when top level block is not disabled', () => {
+			const state = {
+				...baseState,
+				blockEditingModes: new Map( [] ),
+			};
+			expect(
+				isBlockSubtreeDisabled(
+					state,
+					'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
+				)
+			).toBe( false );
 		} );
 
-		describe( 'isBlockSubtreeDisabled', () => {
-			it( 'should return false when top level block is not disabled', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [] ),
-				};
-				expect(
-					isBlockSubtreeDisabled(
-						state,
-						'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
-					)
-				).toBe( false );
-			} );
+		it( 'should return true when top level block is disabled and there are no editing modes within it', () => {
+			const state = {
+				...baseState,
+				blockEditingModes: new Map( [
+					[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
+				] ),
+			};
+			expect(
+				isBlockSubtreeDisabled(
+					state,
+					'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
+				)
+			).toBe( true );
+		} );
 
-			it( 'should return true when top level block is disabled and there are no editing modes within it', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
-					] ),
-				};
-				expect(
-					isBlockSubtreeDisabled(
-						state,
-						'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
-					)
-				).toBe( true );
-			} );
+		it( 'should return true when top level block is disabled via inheritence and there are no editing modes within it', () => {
+			const state = {
+				...baseState,
+				blockEditingModes: new Map( [ [ '', 'disabled' ] ] ),
+			};
+			expect(
+				isBlockSubtreeDisabled(
+					state,
+					'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
+				)
+			).toBe( true );
+		} );
 
-			it( 'should return true when top level block is disabled via inheritence and there are no editing modes within it', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [ [ '', 'disabled' ] ] ),
-				};
-				expect(
-					isBlockSubtreeDisabled(
-						state,
-						'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
-					)
-				).toBe( true );
-			} );
+		it( 'should return true when top level block is disabled and there are disabled editing modes within it', () => {
+			const state = {
+				...baseState,
+				blockEditingModes: new Map( [
+					[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
+					[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'disabled' ],
+				] ),
+			};
+			expect(
+				isBlockSubtreeDisabled(
+					state,
+					'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
+				)
+			).toBe( true );
+		} );
 
-			it( 'should return true when top level block is disabled and there are disabled editing modes within it', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
-						[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'disabled' ],
-					] ),
-				};
-				expect(
-					isBlockSubtreeDisabled(
-						state,
-						'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
-					)
-				).toBe( true );
-			} );
+		it( 'should return false when top level block is disabled and there are non-disabled editing modes within it', () => {
+			const state = {
+				...baseState,
+				blockEditingModes: new Map( [
+					[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
+					[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'default' ],
+				] ),
+			};
+			expect(
+				isBlockSubtreeDisabled(
+					state,
+					'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
+				)
+			).toBe( false );
+		} );
 
-			it( 'should return false when top level block is disabled and there are non-disabled editing modes within it', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ 'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337', 'disabled' ],
-						[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'default' ],
-					] ),
-				};
-				expect(
-					isBlockSubtreeDisabled(
-						state,
-						'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
-					)
-				).toBe( false );
-			} );
-
-			it( 'should return false when top level block is disabled via inheritence and there are non-disabled editing modes within it', () => {
-				const state = {
-					...baseState,
-					blockEditingModes: new Map( [
-						[ '', 'disabled' ],
-						[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'default' ],
-					] ),
-				};
-				expect(
-					isBlockSubtreeDisabled(
-						state,
-						'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
-					)
-				).toBe( false );
-			} );
+		it( 'should return false when top level block is disabled via inheritence and there are non-disabled editing modes within it', () => {
+			const state = {
+				...baseState,
+				blockEditingModes: new Map( [
+					[ '', 'disabled' ],
+					[ 'b3247f75-fd94-4fef-97f9-5bfd162cc416', 'default' ],
+				] ),
+			};
+			expect(
+				isBlockSubtreeDisabled(
+					state,
+					'ef45d5fd-5234-4fd5-ac4f-c3736c7f9337'
+				)
+			).toBe( false );
 		} );
 	} );
